+++ tortoise/backends/base/schema_generator.py
@@ -3,7 +3,6 @@ from hashlib import sha256
 from typing import TYPE_CHECKING, List, Set, Type, cast
 
 from tortoise.exceptions import ConfigurationError
-from tortoise.utils import get_escape_translation_table
 
 if TYPE_CHECKING:  # pragma: nocoverage
     from tortoise.backends.base.client import BaseDBAsyncClient
@@ -36,30 +35,36 @@ class BaseSchemaGenerator:
         self.client = client
 
     def _create_string(
-        self, db_field: str, field_type: str, nullable: str, unique: str, is_pk: bool, comment: str
+        self,
+        db_column: str,
+        field_type: str,
+        nullable: str,
+        unique: str,
+        is_primary_key: bool,
+        comment: str,
     ) -> str:
         # children can override this function to customize their sql queries
 
         return self.FIELD_TEMPLATE.format(
-            name=db_field,
+            name=db_column,
             type=field_type,
             nullable=nullable,
-            unique="" if is_pk else unique,
+            unique="" if is_primary_key else unique,
             comment=comment if self.client.capabilities.inline_comment else "",
-            primary=" PRIMARY KEY" if is_pk else "",
+            primary=" PRIMARY KEY" if is_primary_key else "",
         ).strip()
 
     def _create_fk_string(
         self,
         constraint_name: str,
-        db_field: str,
+        db_column: str,
         table: str,
         field: str,
         on_delete: str,
         comment: str,
     ) -> str:
         return self.FK_TEMPLATE.format(
-            db_field=db_field, table=table, field=field, on_delete=on_delete, comment=comment
+            db_column=db_column, table=table, field=field, on_delete=on_delete, comment=comment
         )
 
     def _table_comment_generator(self, table: str, comment: str) -> str:
@@ -78,11 +83,24 @@ class BaseSchemaGenerator:
         # by default does nothing. If need be, it can be over-written
         return ""
 
+    @classmethod
+    def _get_escape_translation_table(cls) -> List[str]:
+        """escape sequence taken based on definition provided by PostgreSQL and MySQL"""
+        _escape_table = [chr(x) for x in range(128)]
+        _escape_table[0] = "\\0"
+        _escape_table[ord("\\")] = "\\\\"
+        _escape_table[ord("\n")] = "\\n"
+        _escape_table[ord("\r")] = "\\r"
+        _escape_table[ord("\032")] = "\\Z"
+        _escape_table[ord('"')] = '\\"'
+        _escape_table[ord("'")] = "\\'"
+        return _escape_table
+
     def _escape_comment(self, comment: str) -> str:
         # This method provides a default method to escape comment strings as per
         # default standard as applied under mysql like database. This can be
         # overwritten if required to match the database specific escaping.
-        return comment.translate(get_escape_translation_table())
+        return comment.translate(self._get_escape_translation_table())
 
     def _table_generate_extra(self, table: str) -> str:
         return ""
@@ -104,7 +122,7 @@ class BaseSchemaGenerator:
         # NOTE: for compatibility, index name should not be longer than 30
         # characters (Oracle limit).
         # That's why we slice some of the strings here.
-        table_name = model._meta.table
+        table_name = model._meta.db_table
         index_name = "{}_{}_{}_{}".format(
             prefix,
             table_name[:11],
@@ -130,7 +148,7 @@ class BaseSchemaGenerator:
         return self.INDEX_CREATE_TEMPLATE.format(
             exists="IF NOT EXISTS " if safe else "",
             index_name=self._generate_index_name("idx", model, field_names),
-            table_name=model._meta.table,
+            table_name=model._meta.db_table,
             fields=", ".join([self.quote(f) for f in field_names]),
         )
 
@@ -146,11 +164,11 @@ class BaseSchemaGenerator:
         m2m_tables_for_create = []
         references = set()
 
-        for field_name, db_field in model._meta.fields_db_projection.items():
+        for field_name, column_name in model._meta.fields_db_projection.items():
             field_object = model._meta.fields_map[field_name]
             comment = (
                 self._column_comment_generator(
-                    table=model._meta.table, column=db_field, comment=field_object.description
+                    table=model._meta.db_table, column=column_name, comment=field_object.description
                 )
                 if field_object.description
                 else ""
@@ -162,7 +180,9 @@ class BaseSchemaGenerator:
                     if generated_sql:  # pragma: nobranch
                         fields_to_create.append(
                             self.GENERATED_PK_TEMPLATE.format(
-                                field_name=db_field, generated_sql=generated_sql, comment=comment,
+                                field_name=column_name,
+                                generated_sql=generated_sql,
+                                comment=comment,
                             )
                         )
                         continue
@@ -170,11 +190,13 @@ class BaseSchemaGenerator:
             nullable = "NOT NULL" if not field_object.null else ""
             unique = "UNIQUE" if field_object.unique else ""
 
-            if hasattr(field_object, "reference") and field_object.reference:
+            if getattr(field_object, "reference", None):
                 reference = cast("ForeignKeyFieldInstance", field_object.reference)
                 comment = (
                     self._column_comment_generator(
-                        table=model._meta.table, column=db_field, comment=reference.description,
+                        table=model._meta.db_table,
+                        column=column_name,
+                        comment=reference.description,
                     )
                     if reference.description
                     else ""
@@ -185,40 +207,40 @@ class BaseSchemaGenerator:
                     to_field_name = reference.to_field_instance.model_field_name
 
                 field_creation_string = self._create_string(
-                    db_field=db_field,
+                    db_column=column_name,
                     field_type=field_object.get_for_dialect(self.DIALECT, "SQL_TYPE"),
                     nullable=nullable,
                     unique=unique,
-                    is_pk=field_object.pk,
+                    is_primary_key=field_object.pk,
                     comment="",
                 ) + self._create_fk_string(
                     constraint_name=self._generate_fk_name(
-                        model._meta.table,
-                        db_field,
-                        reference.model_class._meta.table,
+                        model._meta.db_table,
+                        column_name,
+                        reference.related_model._meta.db_table,
                         to_field_name,
                     ),
-                    db_field=db_field,
-                    table=reference.model_class._meta.table,
+                    db_column=column_name,
+                    table=reference.related_model._meta.db_table,
                     field=to_field_name,
                     on_delete=reference.on_delete,
                     comment=comment,
                 )
-                references.add(reference.model_class._meta.table)
+                references.add(reference.related_model._meta.db_table)
             else:
                 field_creation_string = self._create_string(
-                    db_field=db_field,
+                    db_column=column_name,
                     field_type=field_object.get_for_dialect(self.DIALECT, "SQL_TYPE"),
                     nullable=nullable,
                     unique=unique,
-                    is_pk=field_object.pk,
+                    is_primary_key=field_object.pk,
                     comment=comment,
                 )
 
             fields_to_create.append(field_creation_string)
 
             if field_object.index and not field_object.pk:
-                fields_with_index.append(db_field)
+                fields_with_index.append(column_name)
 
         if model._meta.unique_together:
             for unique_together_list in model._meta.unique_together:
@@ -253,7 +275,7 @@ class BaseSchemaGenerator:
         table_fields_string = "\n    {}\n".format(",\n    ".join(fields_to_create))
         table_comment = (
             self._table_comment_generator(
-                table=model._meta.table, comment=model._meta.table_description
+                table=model._meta.db_table, comment=model._meta.table_description
             )
             if model._meta.table_description
             else ""
@@ -261,10 +283,10 @@ class BaseSchemaGenerator:
 
         table_create_string = self.TABLE_CREATE_TEMPLATE.format(
             exists="IF NOT EXISTS " if safe else "",
-            table_name=model._meta.table,
+            table_name=model._meta.db_table,
             fields=table_fields_string,
             comment=table_comment,
-            extra=self._table_generate_extra(table=model._meta.table),
+            extra=self._table_generate_extra(table=model._meta.db_table),
         )
 
         table_create_string = "\n".join([table_create_string, *field_indexes_sqls])
@@ -278,14 +300,14 @@ class BaseSchemaGenerator:
             m2m_create_string = self.M2M_TABLE_TEMPLATE.format(
                 exists="IF NOT EXISTS " if safe else "",
                 table_name=field_object.through,
-                backward_table=model._meta.table,
-                forward_table=field_object.model_class._meta.table,
-                backward_field=model._meta.db_pk_field,
-                forward_field=field_object.model_class._meta.db_pk_field,
+                backward_table=model._meta.db_table,
+                forward_table=field_object.related_model._meta.db_table,
+                backward_field=model._meta.db_pk_column,
+                forward_field=field_object.related_model._meta.db_pk_column,
                 backward_key=field_object.backward_key,
                 backward_type=model._meta.pk.get_for_dialect(self.DIALECT, "SQL_TYPE"),
                 forward_key=field_object.forward_key,
-                forward_type=field_object.model_class._meta.pk.get_for_dialect(
+                forward_type=field_object.related_model._meta.pk.get_for_dialect(
                     self.DIALECT, "SQL_TYPE"
                 ),
                 extra=self._table_generate_extra(table=field_object.through),
@@ -299,7 +321,7 @@ class BaseSchemaGenerator:
             m2m_tables_for_create.append(m2m_create_string)
 
         return {
-            "table": model._meta.table,
+            "table": model._meta.db_table,
             "model": model,
             "table_creation_string": table_create_string,
             "references": references,