+++ CHANGELOG.rst
@@ -9,6 +9,34 @@ Changelog
 
 0.16
 ====
+0.16.6
+------
+* Added support for partial models:
+
+  To create a partial model, one can do a ``.only(<fieldnames-as-strings>)`` as part of the QuerySet.
+  This will create model instances that only have those values fetched.
+
+  Persisting changes on the model is allowed only when:
+
+  * All the fields you want to update is specified in ``<model>.save(update_fields=[...])``
+  * You included the Model primary key in the `.only(...)``
+
+  To protect against common mistakes we ensure that errors get raised:
+
+  * If you access a field that is not specified, you will get an ``AttributeError``.
+  * If you do a ``<model>.save()`` a ``IncompleteInstanceError`` will be raised as the model is, as requested, incomplete.
+  * If you do a ``<model>.save(update_fields=[...])`` and you didn't include the primary key in the ``.only(...)``,
+    then ``IncompleteInstanceError`` will be raised indicating that updates can't be done without the primary key being known.
+  * If you do a ``<model>.save(update_fields=[...])`` and one of the fields in ``update_fields`` was not in the ``.only(...)``,
+    then ``IncompleteInstanceError`` as that field is not available to be updated.
+
+0.16.5
+------
+* Moved ``Tortoise.describe_model(<MODEL>, ...)`` to ``<MODEL>.describe(...)``
+* Deprecated ``Tortoise.describe_model()``
+* Fix for ``generate_schemas`` param being ignored in ``tortoise.contrib.quart.register_tortoise``
+* Fix join query with `source_field` param
+
 0.16.4
 ------
 * More consistent escaping of db columns, fixes using SQL reserved keywords as field names with a function.